Extract internal classes
